crm-5: refactor:: add HasPermissions trait to User model
